making a new rail
